Refactor build process into more steps, organized around repos
This is kind of a chunky commit -- sorry.

* Cluster specs' build component now just lists repositories and what to
  do with them.
* The build process checks out each repo, then based on its kind adds
  more tasks to do the appropriate thing
* Tasks common to mulitple repos are broken out and added only once
* Each task's outputs are not modified by subsequent steps, and where
  necessary are "stamped" with the sources so they can be regenerated
  as necessary.

+++ src/build/service.js
@@ -1,5 +1,6 @@
 const _ = require('lodash');
 const util = require('util');
+const assert = require('assert');
 const fs = require('fs');
 const path = require('path');
 const split = require('split');
@@ -9,112 +10,74 @@ const doT = require('dot');
 const {quote} = require('shell-quote');
 const yaml = require('js-yaml');
 const tar = require('tar-fs');
+const copy = require('recursive-copy');
 const {gitClone, dockerRun, dockerPull, dockerImages, dockerBuild, dockerRegistryCheck,
-  dockerPush} = require('./utils');
+  dockerPush, dirStamped, stampDir} = require('./utils');
 
 doT.templateSettings.strip = false;
 const ENTRYPOINT_TEMPLATE = doT.template(fs.readFileSync(path.join(__dirname, 'entrypoint.dot')));
 const DOCKERFILE_TEMPLATE = doT.template(fs.readFileSync(path.join(__dirname, 'dockerfile.dot')));
 
-const serviceTasks = ({baseDir, spec, cfg, name, cmdOptions}) => {
-  const service = _.find(spec.build.services, {name});
+const generateServiceTasks = ({tasks, baseDir, spec, cfg, name, cmdOptions}) => {
+  const repository = _.find(spec.build.repositories, {name});
   const workDir = path.join(baseDir, `service-${name}`);
-  const appDir = path.join(workDir, 'app');
-  const buildpackDir = path.join(workDir, 'buildpack');
-  let stackImage, buildImage;
+  if (!fs.existsSync(workDir)) {
+    fs.mkdirSync(workDir);
+  }
 
-  const tasks = [];
+  assert(repository.service.buildtype === 'heroku-buildpack', 'I only know how to do heroku buildpacks!');
 
-  const writeEntrypointScript = () => {
-    const procfilePath = path.join(appDir, 'Procfile');
-    if (!fs.existsSync(procfilePath)) {
-      throw new Error(`Service ${name} has no Procfile`);
+  const stackImage = `heroku/${repository.service.stack.replace('-', ':')}`;
+  const buildImage = `heroku/${repository.service.stack.replace('-', ':')}-build`;
+  const buildpackUrl = repository.service.buildpack;
+  const buildpackName = buildpackUrl.startsWith('https://github.com/heroku/') ?
+    buildpackUrl.split('/')[4] : buildpackUrl.replace('/', '_');
+
+  // add a task to tasks only if it isn't already there
+  const ensureTask = (task) => {
+    if (!_.find(tasks, {title: task.title})) {
+      tasks.push(task);
     }
-    const Procfile = fs.readFileSync(procfilePath).toString();
-    const procs = Procfile.split('\n').map(line => {
-      if (!line || line.startsWith('#')) {
-        return null;
-      }
-      const parts = /^([^:]+):?\s+(.*)$/.exec(line.trim());
-      if (!parts) {
-        throw new Error(`unexpected line in Procfile: ${line}`);
-      }
-      return {name: parts[1], command: quote([parts[2]])};
-    }).filter(l => l !== null);
-    const entrypoint = ENTRYPOINT_TEMPLATE({procs});
-    fs.writeFileSync(path.join(appDir, 'entrypoint'), entrypoint, {mode: 0o777});
   };
 
-  tasks.push({
-    title: `Service ${name} - Preflight`,
+  ensureTask({
+    title: `Pull Docker Image ${stackImage}`,
     requires: [],
     provides: [
-      `service-${name}-docker-image`, // docker image tag
-      `service-${name}-exact-source`, // exact source URL
-      `service-${name}-image-exists`, // true if the image already exists
-      `service-${name}-image-on-registry`, // true if the image already exists
+      `docker-image-${stackImage}`,
     ],
     run: async (requirements, utils) => {
-      utils.step({title: 'Set Up'});
+      await dockerPull({image: stackImage, utils, baseDir});
+    },
+  });
 
-      if (!fs.existsSync(workDir)) {
-        fs.mkdirSync(workDir);
-      }
+  ensureTask({
+    title: `Pull Docker Image ${buildImage}`,
+    requires: [],
+    provides: [
+      `docker-image-${buildImage}`,
+    ],
+    run: async (requirements, utils) => {
+      await dockerPull({image: buildImage, utils, baseDir});
+    },
+  });
 
-      ['cache', 'env'].forEach(dir => {
-        if (!fs.existsSync(path.join(workDir, dir))) {
-          fs.mkdirSync(path.join(workDir, dir));
-        }
-      });
-
-      utils.step({title: 'Check out Service Repo'});
-
+  ensureTask({
+    title: `Clone ${buildpackName}`,
+    requires: [],
+    provides: [
+      `repo-${buildpackName}`,
+    ],
+    run: async (requirements, utils) => {
+      const repoDir = path.join(baseDir, `repo-${buildpackName}`);
       await gitClone({
-        dir: 'app',
-        url: service.source,
+        dir: repoDir,
+        url: buildpackUrl,
         utils,
-        workDir,
       });
-
-      utils.step({title: 'Read Build Config'});
-
-      // default buildConfig
-      buildConfig = {
-        buildType: 'heroku-buildpack',
-        stack: 'heroku-16',
-        buildpack: 'https://github.com/heroku/heroku-buildpack-nodejs',
-      };
-
-      const buildConfigFile = path.join(appDir, '.build-config.yml');
-      if (fs.existsSync(buildConfigFile)) {
-        const config = yaml.safeLoad(buildConfigFile);
-        Object.assign(buildConfig, config);
-      }
-
-      stackImage = `heroku/${buildConfig.stack.replace('-', ':')}`;
-      buildImage = `heroku/${buildConfig.stack.replace('-', ':')}-build`;
-
-      utils.step({title: 'Check for Existing Image'});
-
-      const [source, ref] = service.source.split('#');
-      const head = (await git(workDir).listRemote([source, ref])).split(/\s+/)[0];
-      const tag = `${cfg.docker.repositoryPrefix}${name}:${head}`;
-
-      // set up to skip other tasks if this tag already exists locally
-      const localDockerImages = await dockerImages({workDir});
-      // TODO: need docker image sha, if it exists (or set it later)
-      const dockerImageExists = localDockerImages.some(image => image.RepoTags && image.RepoTags.indexOf(tag) !== -1);
-
-      utils.step({title: 'Check for Existing Image on Registry'});
-
-      // check whether it's on the registry, too
-      const onRegistry = await dockerRegistryCheck({tag});
 
       return {
-        [`service-${name}-docker-image`]: tag,
-        [`service-${name}-exact-source`]: `${source}#${head}`,
-        [`service-${name}-image-exists`]: dockerImageExists,
-        [`service-${name}-image-on-registry`]: onRegistry,
+        [`repo-${buildpackName}`]: repoDir,
       };
     },
   });
@@ -122,66 +85,95 @@ const serviceTasks = ({baseDir, spec, cfg, name, cmdOptions}) => {
   tasks.push({
     title: `Service ${name} - Compile`,
     requires: [
-      `service-${name}-docker-image`,
-      `service-${name}-image-exists`,
-      `service-${name}-image-on-registry`,
-      `service-${name}-exact-source`,
+      `repo-${name}-dir`,
+      `repo-${name}-exact-source`,
+      `docker-image-${buildImage}`,
+      `repo-${buildpackName}`,
     ],
     provides: [
       `service-${name}-built-app-dir`,
     ],
     run: async (requirements, utils) => {
+      const repoDir = requirements[`repo-${name}-dir`];
+      const appDir = path.join(workDir, 'app');
+      const stampFile = path.join(appDir, '.exact-source');
+      const bpDir = requirements[`repo-${buildpackName}`];
       const provides = {
         [`service-${name}-built-app-dir`]: appDir,
       };
 
-      // bail out early if we can skip this..
-      if (requirements[`service-${name}-image-exists`] || requirements[`service-${name}-image-on-registry`]) {
-        // TODO: need to get app dir from that image..
+      // if we've already built this appDir with this revision, we're done.
+      if (dirStamped({dir: appDir, sources: requirements[`repo-${name}-exact-source`]})) {
         return utils.skip(provides);
+      } else {
+        await rimraf(appDir);
       }
 
-      utils.step({title: 'Check out Buildpack Repo'});
-
-      await gitClone({
-        dir: 'buildpack',
-        url: buildConfig.buildpack,
-        utils,
-        workDir,
-      });
-
-      utils.step({title: 'Pull Stack Image'});
-
-      await dockerPull({image: stackImage, utils, workDir});
-
-      utils.step({title: 'Pull Build Image'});
-
-      await dockerPull({image: buildImage, utils, workDir});
+      utils.step({title: 'Copy Source Repository'});
+      await copy(repoDir, appDir); // TODO: filter out .git
 
       utils.step({title: 'Buildpack Detect'});
 
       await dockerRun({
         image: buildImage,
-        command: ['/workdir/buildpack/bin/detect', '/app'],
-        logfile: 'detect.log',
+        command: ['/buildpack/bin/detect', '/app'],
+        binds: [
+          `${bpDir}:/buildpack`,
+          `${appDir}:/app`,
+        ],
+        logfile: `${workDir}/detect.log`,
         utils,
-        workDir,
+        baseDir,
       });
 
       utils.step({title: 'Buildpack Compile'});
 
+      const envDir = path.join(workDir, 'env');
+      const cacheDir = path.join(workDir, 'cache');
+      if (fs.existsSync(envDir)) {
+        await rimraf(envDir);
+      }
+      [envDir, cacheDir].forEach(dir => {
+        if (!fs.existsSync(dir)) {
+          fs.mkdirSync(dir);
+        }
+      });
+
       await dockerRun({
         image: buildImage,
-        command: ['/workdir/buildpack/bin/compile', '/app', '/workdir/cache', '/workdir/env'],
-        logfile: 'compile.log',
+        command: ['/buildpack/bin/compile', '/app', '/cache', '/env'],
+        binds: [
+          `${bpDir}:/buildpack`,
+          `${appDir}:/app`,
+          `${envDir}:/env`,
+          `${cacheDir}:/cache`,
+        ],
+        logfile: `${workDir}/compile.log`,
         utils,
-        workDir,
+        baseDir,
       });
 
       utils.step({title: 'Create Entrypoint Script'});
 
-      writeEntrypointScript();
+      const procfilePath = path.join(appDir, 'Procfile');
+      if (!fs.existsSync(procfilePath)) {
+        throw new Error(`Service ${name} has no Procfile`);
+      }
+      const Procfile = fs.readFileSync(procfilePath).toString();
+      const procs = Procfile.split('\n').map(line => {
+        if (!line || line.startsWith('#')) {
+          return null;
+        }
+        const parts = /^([^:]+):?\s+(.*)$/.exec(line.trim());
+        if (!parts) {
+          throw new Error(`unexpected line in Procfile: ${line}`);
+        }
+        return {name: parts[1], command: quote([parts[2]])};
+      }).filter(l => l !== null);
+      const entrypoint = ENTRYPOINT_TEMPLATE({procs});
+      fs.writeFileSync(path.join(appDir, 'entrypoint'), entrypoint, {mode: 0o777});
 
+      stampDir({dir: appDir, sources: requirements[`repo-${name}-exact-source`]});
       return provides;
     },
   });
@@ -189,33 +181,43 @@ const serviceTasks = ({baseDir, spec, cfg, name, cmdOptions}) => {
   tasks.push({
     title: `Service ${name} - Build Image`,
     requires: [
-      `service-${name}-docker-image`,
-      `service-${name}-image-exists`,
-      `service-${name}-image-on-registry`,
-      `service-${name}-exact-source`,
+      `repo-${name}-exact-source`,
       `service-${name}-built-app-dir`,
+      `docker-image-${stackImage}`,
     ],
     provides: [
-      `service-${name}-image-built`,
+      `service-${name}-docker-image`, // docker image tag
+      `service-${name}-image-on-registry`, // true if the image already exists on registry
     ],
     run: async (requirements, utils) => {
+      const appDir = requirements[`service-${name}-built-app-dir`];
+      const headRef = requirements[`repo-${name}-exact-source`].split('#')[1];
+      const tag = `${cfg.docker.repositoryPrefix}${name}:${headRef}`;
+
+      utils.step({title: 'Check for Existing Images'});
+
+      const imageLocal = (await dockerImages({baseDir}))
+        .some(image => image.RepoTags && image.RepoTags.indexOf(tag) !== -1);
+      const imageOnRegistry = await dockerRegistryCheck({tag});
+
       const provides = {
-        [`service-${name}-image-built`]: true,
+        [`service-${name}-docker-image`]: tag,
+        [`service-${name}-image-on-registry`]: imageOnRegistry,
       };
 
-      // bail out early if we can skip this..
-      if (requirements[`service-${name}-image-exists`] || requirements[`service-${name}-image-on-registry`]) {
+      // bail out if we can, pulling the image if it's only available remotely
+      if (!imageLocal && imageOnRegistry) {
+        await dockerPull({image: tag, utils, baseDir});
+        return utils.skip(provides);
+      } else if (imageLocal) {
         return utils.skip(provides);
       }
 
-      utils.step({title: 'Build Final Image'});
+      // build a tarfile containing the app directory and Dockerfile
+      utils.step({title: 'Create Docker-Build Tarball'});
 
       const dockerfile = DOCKERFILE_TEMPLATE({stackImage});
       fs.writeFileSync(path.join(workDir, 'Dockerfile'), dockerfile);
-
-      // build a tarfile containing the app directory and Dockerfile, but without
-      // app/.git
-      utils.step({title: 'Creating Docker-Build Tarball'});
 
       const appGitDir = path.join(appDir, '.git');
       const tarball = tar.pack(workDir, {
@@ -225,10 +227,10 @@ const serviceTasks = ({baseDir, spec, cfg, name, cmdOptions}) => {
 
       await dockerBuild({
         tarball,
-        logfile: 'docker-build.log',
+        logfile: `${workDir}/docker-build.log`,
         tag: requirements[`service-${name}-docker-image`],
         utils,
-        workDir,
+        baseDir,
       });
 
       return provides;
@@ -236,30 +238,44 @@ const serviceTasks = ({baseDir, spec, cfg, name, cmdOptions}) => {
   });
 
   tasks.push({
-    title: `Service ${name} - Generate Docs Metadata`,
+    title: `Service ${name} - Generate Docs`,
     requires: [
+      `repo-${name}-exact-source`,
       `service-${name}-built-app-dir`,
+      `docker-image-${stackImage}`,
     ],
     provides: [
-      `service-${name}-docs-dir`,
+      `docs-${name}-dir`,
     ],
     run: async (requirements, utils) => {
-      const docsDir = path.join(workDir, 'docs');
+      const appDir = requirements[`service-${name}-built-app-dir`];
+      const docsDir = path.join(baseDir, `docs-${name}`);
+      const stampFile = path.join(docsDir, '.exact-source');
       const provides = {
-        [`service-${name}-docs-dir`]: docsDir,
+        [`docs-${name}-dir`]: docsDir,
       };
 
-      await rimraf(docsDir);
+      // if we've already built this docsDir with this revision, we're done.
+      if (dirStamped({dir: docsDir, sources: requirements[`repo-${name}-exact-source`]})) {
+        return utils.skip(provides);
+      } else {
+        await rimraf(docsDir);
+      }
 
       await dockerRun({
         image: stackImage,
         command: ['/app/entrypoint', 'write-docs'],
-        env: ['DOCS_OUTPUT_DIR=/workdir/docs'],
-        logfile: 'generate-docs.log',
+        env: [`DOCS_OUTPUT_DIR=/basedir/docs-${name}`],
+        logfile: `${workDir}/generate-docs.log`,
         utils,
-        workDir,
+        binds: [
+          `${appDir}:/app`,
+          `${baseDir}:/basedir`,
+        ],
+        baseDir,
       });
 
+      stampDir({dir: docsDir, sources: requirements[`repo-${name}-exact-source`]});
       return provides;
     },
   });
@@ -268,19 +284,15 @@ const serviceTasks = ({baseDir, spec, cfg, name, cmdOptions}) => {
     title: `Service ${name} - Push Image`,
     requires: [
       `service-${name}-docker-image`,
-      `service-${name}-image-built`,
-      `service-${name}-image-exists`,
       `service-${name}-image-on-registry`,
     ],
     provides: [
     ],
     run: async (requirements, utils) => {
       const tag = requirements[`service-${name}-docker-image`];
-      const provides = {
-      };
 
       if (!cmdOptions.push) {
-        return utils.skip(provides);
+        return utils.skip({});
       }
 
       if (requirements[`service-${name}-image-on-registry`]) {
@@ -288,10 +300,10 @@ const serviceTasks = ({baseDir, spec, cfg, name, cmdOptions}) => {
       }
 
       await dockerPush({
-        logfile: 'docker-push.log',
+        logfile: `${workDir}/docker-push.log`,
         tag,
         utils,
-        workDir,
+        baseDir,
       });
 
       return provides;
@@ -301,4 +313,4 @@ const serviceTasks = ({baseDir, spec, cfg, name, cmdOptions}) => {
   return tasks;
 };
 
-exports.serviceTasks = serviceTasks;
+module.exports = generateServiceTasks;
